requirements.txt
# Discord-Bot/main.py: 1
# Discord-Bot/src/utils.py: 1
# Discord-Bot/commands/announce.py: 1
# Discord-Bot/commands/employments.py: 1
# Discord-Bot/commands/help.py: 1
# Discord-Bot/commands/how_to.py: 1
# Discord-Bot/commands/roles.py: 1
# Discord-Bot/commands/suggestions.py: 1
discord.py[speed]==2.0.1

# Discord-Bot/src/env.py: 2
python_dotenv == 0.21.0